requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.4.1
coveralls==1.2.0
Django==1.11.4
djangorestframework==3.6.4
docopt==0.6.2
hashids==1.2.0
idna==2.6
psycopg2==2.7.3.1
pytz==2017.2
requests==2.18.4
urllib3==1.22
